More type handling

+++ src/snake.cpp
@@ -1,6 +1,6 @@
 #include "snake.h"
 #include <cmath>
-#include <iostream>
+#include <algorithm>
 
 void Snake::Update() {
   SDL_Point prev_cell{
@@ -39,8 +39,8 @@ void Snake::UpdateHead() {
   }
 
   // Wrap the Snake around to the beginning if going off of the screen.
-  head_x = fmod(head_x + grid_width, grid_width);
-  head_y = fmod(head_y + grid_height, grid_height);
+  head_x = (float) fmod(head_x + (float) grid_width, grid_width);
+  head_y = (float) fmod(head_y + (float) grid_height, grid_height);
 }
 
 void Snake::UpdateBody(SDL_Point &current_head_cell, SDL_Point &prev_head_cell) {
